made an abstract class SpawnHandler as parent to handlers of coin and obstacle
