+++ core/src/com/grupp28gdx/game/Coin.java
@@ -1,7 +1,6 @@
 package com.grupp28gdx.game;
 
 import com.badlogic.gdx.physics.box2d.CircleShape;
-import com.badlogic.gdx.physics.box2d.PolygonShape;
 
 public class Coin extends Actor{
     private int posX;
